requirements.txt
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=8.0.0
scikit-learn>=0.24.0
scikit-image
pandas>=1.3.0
tqdm
safetensors
timm
transformers
huggingface_hub
easydict
wandb
catalyst
omegaconf
opencv-python
opencv-contrib-python
tensorflow-datasets
e2cnn
mxnet==1.9.1
pyarrow
facenet-pytorch
accelerate
pyrootutils
ninja
